refactored channels table
