+++ cmd/torq/internal/torqsrv/server.go
@@ -8,9 +8,9 @@ import (
 	"github.com/jmoiron/sqlx"
 	"github.com/lncapital/torq/internal/auth"
 	"github.com/lncapital/torq/internal/channels"
-	"github.com/lncapital/torq/internal/channels/tags"
 	"github.com/lncapital/torq/internal/flow"
 	"github.com/lncapital/torq/internal/settings"
+	"github.com/lncapital/torq/internal/tags"
 	"github.com/lncapital/torq/internal/views"
 	"github.com/ulule/limiter/v3"
 	mgin "github.com/ulule/limiter/v3/drivers/middleware/gin"
